Override toPath() for Polyline.
